SYMBOL REFACTOR
