+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/ComparisonDefinitionHeader.tsx
@@ -6,11 +6,11 @@ import {Button} from '../../../../../Button';
 import {StyledTextField} from '../../StyledTextField';
 import {useCompareEvaluationsState} from './compareEvaluationsContext';
 import {STANDARD_PADDING} from './ecpConstants';
+import {getOrderedCallIds} from './ecpState';
+import {useEvaluationCallDimensions} from './ecpState';
 import {ScoreDimension} from './ecpTypes';
 import {EvaluationComparisonState} from './ecpTypes';
 import {EvaluationDefinition} from './EvaluationDefinition';
-import {getOrderedCallIds} from './evaluationResults';
-import {useEvaluationCallDimensions} from './initialize';
 import {HorizontalBox} from './Layout';
 
 export const ComparisonDefinition: React.FC<{
