+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/compareEvaluationsContext.tsx
@@ -2,10 +2,9 @@ import {Box, LinearProgress} from '@material-ui/core';
 import React, {useMemo} from 'react';
 
 import {WeaveLoader} from '../../../../../../common/components/WeaveLoader';
+import {RangeSelection, useEvaluationComparisonState} from './ecpState';
 import {ScoreDimension} from './ecpTypes';
 import {EvaluationComparisonState} from './ecpTypes';
-import {RangeSelection, useInitialState} from './initialize';
-// import FullScreenLoader from './FullscreenLoader';
 
 const CompareEvaluationsContext = React.createContext<{
   state: EvaluationComparisonState;
@@ -57,7 +56,7 @@ export const CompareEvaluationsProvider: React.FC<{
   selectedInputDigest,
   children,
 }) => {
-  const initialState = useInitialState(
+  const initialState = useEvaluationComparisonState(
     entity,
     project,
     evaluationCallIds,
