+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/ecpTypes.ts
@@ -2,7 +2,7 @@ import {
   TraceCallSchema,
   TraceObjSchema,
 } from '../wfReactInterface/traceServerClient';
-import {RangeSelection} from './initialize';
+import {RangeSelection} from './ecpState';
 
 export type EvaluationComparisonState = {
   data: EvaluationComparisonData;
